2018.11.11 版本0.1.3 网络修复 搜索更改 UI细节优化 BUG修复 作者yui

+++ app/src/main/java/com/riuir/calibur/data/anime/CartoonListInfo.java
@@ -73,7 +73,7 @@ public class CartoonListInfo {
         private int user_id;
         private int bangumi_id;
         private String name;
-        private int part;
+        private float part;
         private int image_count;
         private CartoonListInfoUser user;
         private CartoonListInfoSource source;
@@ -113,7 +113,7 @@ public class CartoonListInfo {
             this.name = name;
         }
 
-        public int getPart() {
+        public float getPart() {
             return part;
         }
